Add proper SSE state machine
The old implementation never closed the stream on the server-side,
instead continuously sending the final message when repolled.

This was probably a very bad idea anyways, but it also turns out that
browsers may still continue reading from streams that are closed on
the client-side, potentially consuming gigabites of data and crashing.