Use tokio notify instead of broadcast channel
The reason for this change is the fact that `Notify` provides a
concrete, owned Future type that can easily be stored in our state
machine, while the `broadcast::Receiver` does not.

This allows us to directly poll the `Notified` future in the
`poll_frame` function, instead of our previous strategy of spawning a
tasks that wakes our future whenever the channel receives a value.

This is important as, in some cases, executors will re-poll a future
that returned `Poll::Pending` before the waker is called. For example,
in #30 the the addition of rustls for HTTPS caused pending futures to
always be polled a second time, triggering a bug.

Also, this change allows us to not require the tokio "rt" feature for
this crate, which is nice.